Add a mechanism to retrieve PipelineComponent by name
For running pipeline recipes, we will need to refer to pipeline components by name.  To facilitate that, we will adds a map between the pipeline component name and its implementing class.

We mandate each PipelineComponent provide a name, and we source that from the same variable that defines the CLI command name.

Test plan: see new test.